[PI-1123]Handle invoice not found during file download (#37)
Thanks to this code when Wfirma responses with a JSON error message after downloading an invoice request, the file response handler will throw a not found exception. This situation is related to the case that the invoice identifier sent to Wfirma is invalid or it is problem with fetching invoice files from their server.