Change logic of VIES company in Poland due to Wfirma tax identifier logic (#39)
https://www.aaronsaray.com/2018/phpunit-dataprovider-before-setup

Proper VIES logic states that when tax identifier is verified by VIES, therefore it has a country prefix. Yet, wfirma does not handle that logic properly. Probably because it is a polish company. That's why when the country is PL and the VIES is verified provided tax identifier does not have a country prefix.