Use proper alpha-2 country code for Greece when printing contractor (#62)
The EU uses `EL` as the code for Greece for VAT administration purposes, but the actual alpha-2 code for the country is `GR`. This may cause inconsistencies depending on whether a particular service (like wFirma) accepts one or the other. This will make sure that when printing Contractor data for the wFirma API, we use the actual alpha-2 code.